extract PG specified logic
