+++ activerecord/lib/active_record/connection_adapters/abstract/connection_pool.rb
@@ -22,7 +22,7 @@ module ActiveRecord
   module ConnectionAdapters
     module AbstractPool # :nodoc:
       def get_schema_cache(connection)
-        self.schema_cache ||= SchemaCache.new(connection)
+        self.schema_cache ||= connection.kind_of?(PostgreSQLAdapter) ? PostgreSQL::SchemaCache.new(connection) : SchemaCache.new(connection)
         schema_cache.connection = connection
         schema_cache
       end
